remove upcast

+++ src/transformers/models/chameleon/modeling_chameleon.py
@@ -1948,12 +1948,11 @@ class ChameleonForConditionalGeneration(ChameleonPreTrainedModel, GenerationMixi
 
         hidden_states = outputs[0]
         logits = self.lm_head(hidden_states)
-        logits = logits.float()
 
         loss = None
         if labels is not None:
             # Upcast to float if we need to compute the loss to avoid potential precision issues
-            logits = logits.float()
+            # logits = logits.float()
             # Shift so that tokens < n predict n
             shift_logits = logits[..., :-1, :].contiguous()
             shift_labels = labels[..., 1:].contiguous()
